Extract desktop app health checks
